Visual improvements and Mobile support
Now, the user can change the size of the table reactioning to the mobile reaction, it is detected automatically by user information too.
Now, every command that doesn't exist will be deleted and a message will be sended to the author of the message to notice him that this command doesn't exist.